pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fano"
version = "0.1.0"
description = "Multi-component platform for mathematical exploration and documentation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # HTTP clients
    "aiohttp>=3.9.0",
    "requests>=2.31.0",

    # Web frameworks
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "flask>=3.0.0",

    # Data validation
    "pydantic>=2.5.0",

    # LLM APIs
    "anthropic>=0.39.0",

    # Browser automation
    "playwright>=1.40.0",

    # Database
    "sqlalchemy>=2.0.0",

    # Configuration
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "rich>=13.0.0",
    "asyncio-throttle>=1.0.0",
    "markdown>=3.5.0",
    "python-dateutil>=2.8.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "shared*",
    "llm*",
    "pool*",
    "explorer*",
    "documenter*",
    "control*",
]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.html", "*.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["llm/tests", "pool/tests"]
# Handle module name conflicts between pool.src.models and llm.src.models
# by running tests from component directories
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
